chore: reorganize e2e helpers
Split e2e/helpers into backend/ and ui/ subdirectories, with backend/index.ts as the public surface (btc/ln/mint/orchard/tap). Move shared type definitions out to e2e/types/ and switch all imports to the @e2e/* alias via a new e2e/tsconfig.json wired through Playwright's `tsconfig` option. No behavior change — pure restructuring of the e2e test infra introduced earlier on this branch.
