+++ e2e/specs/bitcoin-general-info.spec.ts
@@ -21,8 +21,8 @@
 
 import {test, expect, type Locator, type Page} from '@playwright/test';
 
-import {getConfig} from '../helpers/config';
-import {btc} from '../helpers/backend';
+import {getConfig} from '@e2e/helpers/config';
+import {btc} from '@e2e/helpers/backend';
 
 /** Read the numeric/text value displayed under a metric label in the card. */
 async function metricValue(card: Locator, label: string): Promise<string> {
