+++ e2e/specs/lightning-general-channel-summary.spec.ts
@@ -32,8 +32,8 @@
 
 import {test, expect, type Locator, type Page} from '@playwright/test';
 
-import {getConfig} from '../helpers/config';
-import {ln} from '../helpers/backend';
+import {getConfig} from '@e2e/helpers/config';
+import {ln} from '@e2e/helpers/backend';
 
 async function openSummary(page: Page): Promise<Locator> {
 	const card = page.locator('orc-lightning-general-channel-summary');
